Did changes as per rmehta ( request.__getattr, login params and breaking of session.py). Will now have to get things working
